add type=0 for byosed; fixed a float type mismatch bug in extinction

+++ salt3/training/TrainSALT.py
@@ -114,7 +114,7 @@ def specflux(obsphase,obswave,m0phase,m0wave,m0flux,m1flux,colorlaw,z,x0,x1,c,mw
 	intwave = interp1d(np.unique(m0wave)*(1+z),modelflux_phase,kind='nearest',bounds_error=False,fill_value="extrapolate")
 	modelflux_wave = intwave(obswave)
 	modelflux_wave = x0*(m0interp + x1*m1interp)
-	mwextcurve = 10**(-0.4*extinction.fitzpatrick99(obswave,mwebv*3.1))
+	mwextcurve = 10**(-0.4*extinction.fitzpatrick99(np.float64(obswave),mwebv*3.1))
 	modelflux_wave *= mwextcurve
 
 	return modelflux_wave
